+++ src/command_modules/azure-cli-eventgrid/HISTORY.rst
@@ -2,6 +2,15 @@
 
 Release History
 ===============
+0.2.4
++++++
+* `domain create/update/delete/get/list`: Added new commands for the domain CRUD operations.
+* `domain topic create/update/delete/get/list`: Added new commands for the domain topics CRUD operations.
+* `topic list`: Added --odata-query for filtering the results using OData syntax.
+* `event-subscription list`: Added --odata-query for filtering the results using OData syntax.
+* `event-subscription create/update`: Added servicebusqueue as new values for the `--endpoint-type` parameter.
+* `event-subscription create/update`: Disallow usage of All for --included-event-types.
+
 0.2.3
 +++++
 * Minor fixes.
@@ -13,7 +22,7 @@ Release History
 0.2.1
 +++++
 * `event-subscription create/update`: Added `--deadletter-endpoint` parameter.
-* 'event-subscription create/update`: Added storagequeue and hybridconnection as new values for the `--endpoint-type` parameter.
+* `event-subscription create/update`: Added storagequeue and hybridconnection as new values for the `--endpoint-type` parameter.
 * `event-subscription create`: Added `--max-delivery-attempts` and `--event-ttl` parameters to specify the retry policy for events.
 * `event-subscription create/update`: Added a warning message for manual handshake validation when Webhook as destination is used for an event subscription.
 * Added source-resource-id parameter for all event subscription related commands and mark all other source resource related parameters as deprecated.